requirements.txt
openai>=1.30.0
pyannote.audio>=3.1.0
python-dotenv>=1.0.0
torch>=2.0.0
torchaudio>=2.0.0

# Optional: only needed for --local mode (fully offline transcription)
# openai-whisper>=20231117
